[DNM] R6Sensitivity: Write out integers
Siege keeps rejecting my script changes for 2 things not sure why maybe this fixes it?

+++ implementations/R6Sensitivity.py
@@ -131,4 +131,4 @@ class R6Sensitivity(BaseSensitivity):
         scopes = ["Global", "1x", "1xHalf", "2x", "2xHalf", "3x", "4x", "12x"]
         ratios = [1, 0.67, 1, 1, 1, 1, 1, 1]
         for index, scope in enumerate(scopes):
-            BaseSensitivity._replace_singular_config(f"{prefix}{scope}", str(ratios[index] * 100), content)
+            BaseSensitivity._replace_singular_config(f"{prefix}{scope}", str(int(ratios[index] * 100)), content)
